requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.46.0
beautifulsoup4==4.13.4
blinker==1.9.0
bottle==0.13.4
bottle-websocket==0.2.9
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
dash==3.1.1
docx==0.2.4
Eel==0.18.2
et_xmlfile==2.0.0
Flask==3.1.1
fonttools==4.58.5
future==1.0.0
gevent==25.5.1
gevent-websocket==0.10.1
google==3.0.0
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
googleAPI==0.1.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
httplib2==0.22.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.45.0
nest-asyncio==1.6.0
numpy==2.3.0
oauth2client==4.1.3
oauthlib==3.3.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pefile==2023.2.7
pillow==11.2.1
plotly==6.2.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv>=1.0.0
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.32.4
requests-oauthlib==2.0.0
retrying==1.4.0
rsa==4.9.1
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
zope.event==5.1.1
zope.interface==7.2
